requirements.txt
click
requests
progressbar2
